Create BloggingService
In this implementation, we have a users array that stores the registered users for the social blogging app. The posts array stores the posts created by the users. Each post has an associated author user, a createdAt date, an updatedAt date, a number of likes, and an array of comments.

The comments array stores the comments created by the users. Each comment has an associated author user, a post that it is associated with, a createdAt date, an updatedAt date, and a number of likes.

The createPost method creates a new post and adds it to the posts array. The getPost method retrieves a post with the specified ID. The updatePost method updates an existing post with the specified changes. The deletePost method deletes a post with the specified ID.

The createComment method creates a new comment and adds it to the comments array. It also adds the new comment to the comments array of the associated post. The getComment method retrieves a comment with the specified ID. The updateComment method updates an existing comment with the specified changes. The deleteComment method deletes a comment with the specified ID. It also removes the comment from the comments array of the associated post.

Note that this implementation uses a simple in-memory database with mock data for demonstration purposes only. In a real application, you would likely use a database system such as MySQL or MongoDB to store the data.

by Kidus berhanu

The findPostsByAuthor function returns an array of posts written by a particular author. The findPostsByKeyword function returns an array of posts that contain a particular keyword in the title or content.

The findCommentsByAuthor function returns an array of comments written by a particular author. The findCommentsByPost function returns an array of comments associated with a particular post.

The addPostLike function adds a user's ID to the likes array of a post. The removePostLike function removes a user's ID from the likes array of a post.

The addCommentLike function adds a user's ID to the likes array of a comment. The removeCommentLike function removes a user's ID from the likes array of a comment.

These functions can be useful for implementing features such as search, filtering, and user interactions like liking and disliking posts and comments