Create enhance performance .js
optimize the r site's code and images, minimize HTTP requests, and enable caching, you can use various techniques and tools available in TypeScript. 
 using  the Express framework to set up a server-side application that implements several techniques for optimizing the performance of your site:

Compression is enabled using the compression package to minimize the size of response data and reduce the amount of data that needs to be transmitted.
The helmet middleware is used to set several HTTP headers that improve security and can also improve performance by reducing the amount of data that needs to be transmitted.
Request logging is enabled using the morgan package to help with debugging and performance analysis.
Static files are served from a cache using the express.static middleware. This reduces the number of HTTP requests that need to be made to retrieve these files and can also speed up the delivery of the files by serving them directly from the cache.
A rate limiter is implemented using the rate-limiter-flexible package to prevent abuse and limit the number of requests made to the server.